create a component for popup
